Refactored test data generation.
The generate_test_data function now looks for the testdata.json file in the current working directory and loads the
json into database objects. This relies on the keys of the json objects matching with the column names for the
corresponding schemas. Currently it's only loading File and Job objects, but I'll add others as their models are built.