Work in progress converting to just use jobid.
I realized that the whole purpose of having references to the file in the jobs table was
so that you could just have the job_id you want to process and do everything. In the middle
of refactoring to make that work as well as updating the status of the job as it's in progress.

I also want to abstract the session management stuff since each of these functions is going
to have a very similar session lifecycle. Also, I need to reference some of the sqlalchemy
session management documentation, since they have some opinions about session lifecycles.